pyproject.toml
[project]
name = "dipt-stickers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2024.10",
    "django>=5.1.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "six>=1.16.0",
]
